+++ src/packets/protocol.ts
@@ -13,7 +13,7 @@ import { FirmwareProgressEvent, PacketReceivedEvent, PrintProgressEvent } from "
 import { PrintTaskName, printTasks } from "../print_tasks";
 import { AbstractPrintTask, PrintOptions } from "../print_tasks/AbstractPrintTask";
 import { Validators, Utils } from "../utils";
-import { ConnectNegotiateResult, HeartbeatData, HeartbeatPrinterInfoData, PrinterInfo, PrintError, PrinterStatusData, PrintStatus, RfidInfo } from "./dto";
+import { ConnectNegotiateResult, HeartbeatData, HeartbeatPrinterInfoData, PaperInfo, PrinterInfo, PrintError, PrinterStatusData, PrintStatus, RfidInfo } from "./dto";
 import { NiimbotCrc32Packet, NiimbotPacket } from "./packet";
 import { PacketGenerator } from "./packet_generator";
 import CRC32 from "crc-32";
@@ -427,5 +427,10 @@ export class NiimbotProtocol {
 
   public async setPrinterTime(value: Date = new Date()): Promise<void> {
     await this.send(PacketGenerator.setPrinterTime(value));
+  }
+
+  public async getPaperInfo(): Promise<PaperInfo> {
+    const response = await this.send(PacketGenerator.getPaperInfo());
+    return PacketParser.parsePaperInfoResponse(response);
   }
 }